Версия "последняя надежда".3

+++ src/lesson2/task2/Logical.kt
@@ -1,4 +1,5 @@
 @file:Suppress("UNUSED_PARAMETER")
+
 package lesson2.task2
 
 import lesson1.task1.sqr
@@ -28,9 +29,10 @@ val isNumberHappy = fun(number: Int): Boolean =
  * Считать, что ферзи не могут загораживать друг друга.
  */
 fun queenThreatens(x1: Int, y1: Int, x2: Int, y2: Int): Boolean {
-    val a = y1 - y2
-    val b = x1 - x2
-    return (a == 0) || (b == 0) || (a / b == 1) || (a / b == -1)
+    val a = (y1 - y2).toDouble()
+    val b = (x1 - x2).toDouble()
+    return (a == 0.0) || (b == 0.0) ||
+            (a / b == 1.0) || (a / b == -1.0)
 }
